fix: only create patch for 'fields' property of entities (#916)
A customer had an issue where `contentful-merge create` would error due
to the fact that the `displayField` property of the content type in the
target environment was different. However, we are not interested in
changes to top-level properties of the content type that are not related
to `fields`.

Therefore I changed the filter to explicitly create a patch for the
changes in the `fields` property