fix: extract links from rich text field [FUS-556] (#1050)
The Merge App performs a sorting during `apply` command that puts all
entries with links to other entries at the end of the list to ensure
that all the links are published before the main entry. This however was
not the case for rich text fields, which caused issues with merging.

This PR introduces a function that detects links to other entries in
rich text field and sorts them down accordingly.